refactor: extracted business logic in forms
